Tile Map; Binary Collision Implemented tile mapping support. Implemented binary collision support within the tile map. Updated the example redsquare update function to include some new physics. Revamped the input system and refactored code.

+++ AsciiEngine/Shapes.h
@@ -21,10 +21,10 @@ typedef struct _AE_CIRCLE
 typedef struct _AE_RECT
 {
   VECTOR2D center_;
-  int width_;
-  int height_;
-  int halfWidth_;
-  int halfHeight_;
+  float width_;
+  float height_;
+  float halfWidth_;
+  float halfHeight_;
 } AE_RECT;
 
 #endif  SHAPESH
